updates opulo feeder code to work with new encoder motors, along with tracking encoder movement throughout runtime as opposed to relative movements every move command to increase accuracy

+++ photon/src/Feeder.h
@@ -16,7 +16,7 @@ class Feeder {
         };
 
         virtual bool init() = 0;
-        virtual FeedResult feedDistance(uint8_t tenths_mm, bool forward) = 0;
+        virtual FeedResult feedDistance(uint16_t tenths_mm, bool forward) = 0;
 
 };
